configure workers per table
for the tablets model, each table has independent generations and
independent tasks.

For the vnode-based model, the Master transport stores a
currentGenerationId. For the tablets model, we need to store a current
generation per table.

We add a new configureWorkers method that allows to configure workers
for a specific table, and will be used in the tablets model. This method
updates the current generation id for that table, and updates the worker
thread with the tasks for the table by adding them dynamically - not
affecting the tasks of other tables.