reread with max consumed timestamp
When tasks are completed for some generation, find the max consumed
timestamp, because it might happen that some task consumed a row with a
timestamp greater than the generation's end timestamp.

In that case, we need to revisit all streams and read them up to the
last consumed timestamp, and then start reading the next generation from
this timestamp.